Rename surveys.py to survey.py to be consistent with other names (singular)
