+++ hips/tiles/__init__.py
@@ -1,5 +1,5 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 """Classes and functions to manage HiPS tiles."""
 from .tile import *
-from .surveys import *
+from .survey import *
 from .allsky import *
