+++ hips/tiles/survey.py
