+++ hips/tiles/tests/test_survey.py
@@ -5,7 +5,7 @@ from astropy.utils.data import get_pkg_data_filename
 from astropy.tests.helper import remote_data
 from ...utils.testing import get_hips_extra_file, requires_hips_extra
 from ..tile import HipsTileMeta
-from ..surveys import HipsSurveyProperties, HipsSurveyPropertiesList
+from ..survey import HipsSurveyProperties, HipsSurveyPropertiesList
 
 
 class TestHipsSurveyProperties:
